enqueue: prevent batching for ops from different p2pEpoch
When there are more than 1 op enqueued on per peer, the scheduling gets
flattened.
In that case, we must prevent ops from different epoch to be merge
together as this does not guarantee consistent merging decisions, and
leads to a hang.

To avoid the hang, we make sure to not merge operations from different
epochs.

BUG 5666112