Add global cross NIC support flag for GIN support check
Check that cross NIC is supported by the entire group to determine if
the group supports GIN.

Strictly speaking, whether a process in a group supports cross NIC
should not affect other processes in the group. However, we need to
develop a global view of cross NIC support because each process is
determining the entire groups GIN support individually using
information provided through the allgathers.

Signed-off-by: Ahsan Pervaiz <apervaiz@nvidia.com>