Fix QP retrieval logic to use local devices count
This fix/change addresses two bugs:

1. In the function `ncclIbCommBaseGetQpByIndex`:

During the retrieval of QPs by their index relative to a device when the
number of local devices is different than the number of remote devices within a
communicator. Only in this asymmetric case comm->base.vProps.ndevs and
base::nDataQps are different, and the previous logic for retrieving QPs by
index relative to a device was incorrect.

2. In the function `ncclIbCommBaseGetQpByQpNum`:

The loop that iterates over the QP indices relative to a device was incorrectly
using base::nDataQps instead of the correct value of base::nqps divided by
base::vProps.ndevs. This could lead to a result which does not find the QP even
if it exists. For example, when there are 2 local devices, 2 remote devices and
`ncclParamIbQpsPerConn()=4`, the total number of QPs is 8, `nDataQps` is 2
but the maximal index relative to a device is 3 (8/2 - 1). The previous
logic would only iterate from 0 to 1, thus missing the QPs with indices from
2 to 3 relative to the device.

The fix modifies the logic in both functions to correctly reflect the
organization of QPs in the `base::qps[]` array.

The QPs are created and organized in the base::qps[] array as follows:

The logic that sets the different parameters that determine the total number of
QPs in the `base::qps[]` array can be written as follows:

```c++
comm->base.nDataQps = std::max(comm->base.vProps.ndevs, remoteVProps.ndevs);
comm->base.nqps     = ncclParamIbQpsPerConn() * comm->base.nDataQps;
```

Then, when QPs are created, they are created in "stipes" across the local
devices. Meaning that first QP is created on device 0, second on device 1,
and so on. This can be seen in the functions where the receiver and sender
create their QPs:

```c++
for (int qpIndex = 0; qpIndex < nqps; qpIndex++) {
    uint devIndex = qpIndex % comm->base.vProps.ndevs;
    // ...
}
```

For example, given `nDataQps = max(local=4,remote=6)` and
`ncclParamIbQpsPerConn()=1` the QPs array will have total of 6 QPs and they
will be opened on the 4 local devices as follows:

* qp[0] -> dev0
* qp[1] -> dev1
* qp[2] -> dev2
* qp[3] -> dev3
* qp[4] -> dev0
* qp[5] -> dev1

And if `nDataQps = max(local=4,remote=6)` and `ncclParamIbQpsPerConn()=2` the
total number of QPs will be 12 and they will be opened on the 4 local devices
as follows:

* qp[0] -> dev0
* qp[1] -> dev1
* qp[2] -> dev2
* qp[3] -> dev3
* qp[4] -> dev0
* qp[5] -> dev1
* qp[6] -> dev2
* qp[7] -> dev3
* qp[8] -> dev0
* qp[9] -> dev1
* qp[10] -> dev2
* qp[11] -> dev3

It's possible to think of the QPs being created in groups. The **size** of each
group of QPs equals to number of local devices and the number of groups is
determined by the ceiling of total number of QPs divided by the number of local
devices.

The above construction leads to the following:
1. Every group has a most QPs equal to the number of local devices.
2. The last group may have less QPs than the number of local devices.
3. Within every group, every QP is created on a different local device.

Therefore, when retrieving a QP by its index relative to a device, the logic
should be modified to reflect the above construction.

```c++
// qpIndex is the index relative to a device.
static inline ncclResult_t ncclIbCommBaseGetQpByIndex(struct ncclIbNetCommBase*
                           commBase, int devIndex, int qpIndex, ncclIbQp** qp) {
  assert(devIndex >= 0 && devIndex < commBase->vProps.ndevs);
  *qp = &(commBase->qps[commBase->vProps.ndevs*qpIndex + devIndex]);
  return ncclSuccess;
}
```

For example, given `localDevs=4`, `remoteDevs=6` and
`ncclParamIbQpsPerConn()=2`, then the QPs are created as follows:
* qp[0] -> dev0
* qp[1] -> dev1
* qp[2] -> dev2
* qp[3] -> dev3
* qp[4] -> dev0
* qp[5] -> dev1
* qp[6] -> dev2
* qp[7] -> dev3
* qp[8] -> dev0
* qp[9] -> dev1
* qp[10] -> dev2
* qp[11] -> dev3

Then:
* Input of `qpIndex=0` and `devIndex=2` will retrieve `qp[2]`
    * Which is the 1st QP (index 0) relative to device 2.
* Input of `qpIndex=2` and `devIndex=0` will retrieve `qp[8]`
    * Which is the 3rd QP (index 2) relative to device 0.
* Input of `qpIndex=1` and `devIndex=3` will retrieve `qp[7]`
    * Which is the 2nd QP (index 1) relative to device 3.

Another example, given `localDevs=4`, `remoteDevs=6` and
`ncclParamIbQpsPerConn()=1`, then the QPs are created as follows:
* qp[0] -> dev0
* qp[1] -> dev1
* qp[2] -> dev2
* qp[3] -> dev3
* qp[4] -> dev0
* qp[5] -> dev1

Then:
* Input of `qpIndex=1` and `devIndex=1` will retrieve `qp[5]`
    * Which is the 2nd QP (index 1) relative to device 1.
* Input of `qpIndex=1` and `devIndex=2` will result in an error!
    * Because there is no 2nd QP (index 1) relative to device 2.